Add Flash Card Mode, which includes pagination support
Adding a new mode, Flash Card Mode.

In normal mode, the page is populated with randomly generated chords, using the selected modifiers. Repeats are allowed in this mode.

In Flash Card mode, a single chord is generated for each note and chord modifier pair. As the number can be higher than what can be displayed on the page (currently 16), pagination support has been added to allow pagination over the collection. The order of the chords is random, but each is guaranteed to be shown exactly once.